requirements.txt
onnxruntime @ ./onnxruntime-1.22.0-cp311-cp311-linux_aarch64.whl
opencv-python
torch
torchvision
numpy==1.24
Pillow
